Separate view logic from VehicleMode
