+++ CommonCoreLegacy/src/main/java/skedgo/tripkit/routing/VehicleMode.java
@@ -1,8 +1,6 @@
 package skedgo.tripkit.routing;
 
 import android.content.res.Resources;
-import android.graphics.Color;
-import android.graphics.LightingColorFilter;
 import android.graphics.drawable.Drawable;
 import android.support.annotation.DrawableRes;
 import android.support.annotation.NonNull;
@@ -91,19 +89,6 @@ public enum VehicleMode {
 
   public static VehicleMode[] getPublicTransportModes() {
     return new VehicleMode[] {FERRY, TRAIN, MONORAIL, SUBWAY, BUS, TRAM, CABLECAR};
-  }
-
-  /**
-   * FIXME: This is view-related logic.
-   */
-  public static Drawable createLightDrawable(@NonNull Resources resources, int iconRes) {
-    Drawable drawable = resources.getDrawable(iconRes);
-    if (drawable != null) {
-      drawable = drawable.mutate();
-      drawable.setColorFilter(new LightingColorFilter(Color.parseColor("#FF000000"), 0xffffffff));
-    }
-
-    return drawable;
   }
 
   public boolean isPublicTransport() {
